refactor(reader): merge NavBar and gesture footer drawer
Unify the persistent bottom bar and the swipe-up gesture drawer into a
single coherent ReaderBottomBar. The drawer now only carries
complementary controls (page grid, font size, share, favorite); the
bar keeps the persistent controls (prev/next, page counter, TTS, speed
reader, typography toggle), eliminating the prior duplication between
the two surfaces. All existing data-testids are preserved.
