+++ CHANGELOG.md
@@ -10,6 +10,16 @@ Release populated from the matching section below.
 
 ## [Unreleased]
 
+### Changed
+- Merged the bottom `NavBar` and the gesture footer drawer into a single
+  `ReaderBottomBar` component. The persistent bar and the swipe-up drawer are
+  now one coherent surface; the drawer drops controls already shown in the bar
+  (prev/next, page counter, TTS) and keeps complementary ones (page grid, font
+  size, share, favorite). All existing `data-testid`s are preserved
+  (`nav-bar`, `prev-page`, `next-page`, `page-counter`, `tts-toggle`,
+  `tts-stop`, `speed-reader-toggle`, `gesture-footer-drawer-grid`,
+  `gesture-page-*`).
+
 ### Added
 - **Senior-targeted features**: two new feature flags, both default `off`.
   - `text-to-speech`: Web Speech API narration for any story. Adds
